requirements.txt
django-haystack
django-tastypie
elasticsearch
django-extensions
